requirements.txt
scapy~=2.5.0
Faker~=30.3.0
psycopg2-binary~=2.9.10
Flask~=3.1.0
waitress~=3.0.2
Werkzeug~=3.1.3
requests~=2.32.3
DateTime~=5.5
pillow~=10.4.0